# packages/pstack/api/openapi.yaml
openapi: 3.0.3
...
info:
  title: pstack control plane
  version: 0.34.0
  description: |
    The HTTP API of a pstack host. Every operation here is reachable as `pstack api …`.

    Authentication is `Authorization: Bearer <token>` — either `PSTACK_TOKEN` (the machine
    credential, which outranks every role) or a personal token minted at `/api/tokens`. The CLI
    sends whichever `PSTACK_TOKEN` holds and talks to `PSTACK_API_URL`, which has no default.
servers:
  - url: '{apiUrl}'
    variables:
      apiUrl:
        default: https://api.preview.example.com
        description: PSTACK_API_URL — the API's own hostname, `api.<domain>` on a host stood up by `pstack init`.
...
paths:
...
  /api/probe/{id}:
    parameters: [{ $ref: '#/components/parameters/DeploymentId' }]
    get:
      tags: [host]
      operationId: probeDeployment
      x-cli-name: probe
      summary: Is this preview serving? No token, no body — the answer is the status code and `x-pstack-probe`.
      parameters:
        - { name: service, in: query, schema: { type: string }, description: Which service, on a stack that publishes several. }
      responses:
        '200': { description: The upstream's own status. There is never a body. }

  # ── portable configuration ──────────────────────────────────────────────────────────────────────
...
components:
  parameters:
    DeploymentId:
      name: id
      in: path
      required: true
      schema: { type: string }
      description: The registry id, e.g. `pr-123`.
...